refactor(booking): extract shared booking mutation workflow
